changed price type to BigDecimal

+++ src/main/model/Amazon.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Amazon extends Stock {
-    public Amazon(String company, int price) {
+    public Amazon(String company, BigDecimal price) {
         super(company, price);
     }
 }
