+++ src/main/model/Apple.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Apple extends Stock {
-    public Apple(String company, int price) {
+    public Apple(String company, BigDecimal price) {
         super(company, price);
     }
 }
