+++ src/main/model/Google.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Google extends Stock {
-    public Google(String company, int price) {
+    public Google(String company, BigDecimal price) {
         super(company, price);
     }
 }
