+++ src/main/model/Nvidia.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Nvidia extends Stock {
-    public Nvidia(String company, int price) {
+    public Nvidia(String company, BigDecimal price) {
         super(company, price);
     }
 }
