+++ src/main/model/Tesla.java
@@ -1,7 +1,9 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Tesla extends Stock {
-    public Tesla(String company, int price) {
+    public Tesla(String company, BigDecimal price) {
         super(company, price);
     }
 }
